stable and more efficient packet reading

+++ MinecraftBot/varint.cpp
@@ -4,15 +4,15 @@
  * */
 #include "varint.h"
 
-uint32_t Varint::decode_unsigned_varint( const uint8_t *const data, int &decoded_bytes )
+uint64_t Varint::decode_unsigned_varint( const uint8_t *const data, int &decoded_bytes )
 {
     int i = 0;
-    uint32_t decoded_value = 0;
+    uint64_t decoded_value = 0;
     int shift_amount = 0;
 
     do
     {
-        decoded_value |= (uint32_t)(data[i] & 0x7F) << shift_amount;
+        decoded_value |= (uint64_t)(data[i] & 0x7F) << shift_amount;
         shift_amount += 7;
     } while ( (data[i++] & 0x80) != 0 );
 
@@ -22,7 +22,7 @@ uint32_t Varint::decode_unsigned_varint( const uint8_t *const data, int &decoded
 
 int64_t Varint::decode_signed_varint( const uint8_t *const data, int &decoded_bytes )
 {
-    uint32_t unsigned_value = decode_unsigned_varint(data, decoded_bytes);
+    uint64_t unsigned_value = decode_unsigned_varint(data, decoded_bytes);
     return (int64_t)( unsigned_value & 1 ? ~(unsigned_value >> 1)
                                          :  (unsigned_value >> 1) );
 }
@@ -30,7 +30,7 @@ int64_t Varint::decode_signed_varint( const uint8_t *const data, int &decoded_by
 
 // Encode an unsigned 64-bit varint.  Returns number of encoded bytes.
 // 'buffer' must have room for up to 10 bytes.
-int Varint::encode_unsigned_varint(uint8_t *const buffer, uint32_t value)
+int Varint::encode_unsigned_varint(uint8_t *const buffer, uint64_t value)
 {
     int encoded = 0;
 
@@ -55,9 +55,9 @@ int Varint::encode_unsigned_varint(uint8_t *const buffer, uint32_t value)
 // encoder.  'buffer' must have room for up to 10 bytes.
 int Varint::encode_signed_varint(uint8_t *const buffer, int64_t value)
 {
-    uint32_t uvalue;
+    uint64_t uvalue;
 
-    uvalue = uint32_t( value < 0 ? ~(value << 1) : (value << 1) );
+    uvalue = uint64_t( value < 0 ? ~(value << 1) : (value << 1) );
 
     return encode_unsigned_varint( buffer, uvalue );
 }
